atualização banco de dados

+++ backend/src/main/java/com/senacproject/demo/itens/Itens.java
@@ -17,12 +17,12 @@ public class Itens {
     private Long id;
     private String title;
     private String image;
-    private  Integer price;
+    private Double price;
 
     public Itens(ItensRequestDTO data){
         this.title = data.title();
         this.image = data.image();
-        this.price = data.price();
+        this.price = Double.valueOf(data.price());
     }
 
 }
